Making further changes based on PR comments. Now we use an AbstractDBHelper to help remove overlapping code that was present in the PostgresHelper and SQLServerHelper
