+++ libs/messaging/db-message-bus-impl/src/main/kotlin/net/corda/messagebus/db/persistence/DbUtils.kt
@@ -1,8 +1,7 @@
 package net.corda.messagebus.db.persistence
 
-import net.corda.db.core.DBBaseDataSourceFactory
-import net.corda.db.core.HikariDataSourceFactory
 import net.corda.db.core.InMemoryDataSourceFactory
+import net.corda.db.core.createDataSource
 import net.corda.db.schema.DbSchema.DB_MESSAGE_BUS
 import net.corda.orm.DbEntityManagerConfiguration
 import net.corda.orm.EntityManagerFactoryFactory
@@ -22,7 +21,7 @@ internal fun EntityManagerFactoryFactory.create(
     val dbSource = if (jdbcUrl == null || jdbcUrl.contains("hsqldb", ignoreCase = true)) {
         InMemoryDataSourceFactory().create(DB_MESSAGE_BUS)
     } else {
-        DBBaseDataSourceFactory(HikariDataSourceFactory()).create(
+        createDataSource(
             "org.postgresql.Driver",
             jdbcUrl,
             jdbcUser,
